requirements.txt
# ML
dm-haiku
jmp
chex
optax
einops

# Data
torch
torchvision
tqdm
Pillow
lmdb

# Other
click
pyyaml
pydantic
matplotlib
pandas

# Dev-Ops
mypy
flake8
pytest
isort
